+++ src/main/java/cz/muni/fi/pv260/team4/tictactoe/validators/WinningSequenceLengthValidator.java
@@ -6,9 +6,9 @@ import lombok.AllArgsConstructor;
 import java.util.Optional;
 
 @AllArgsConstructor
-public final class WinningSequenceLengthValidator implements InputValidator<Long> {
-    private final Long boardWidth;
-    private final Long boardHeight;
+public final class WinningSequenceLengthValidator implements InputValidator<Integer> {
+    private final Integer boardWidth;
+    private final Integer boardHeight;
 
     /**
      * Validate the length of the winning sequence.
@@ -17,7 +17,7 @@ public final class WinningSequenceLengthValidator implements InputValidator<Long
      * @return Optional containing error message if validation fails, otherwise empty
      */
     @Override
-    public Optional<String> validate(final Long lengthOfTheWinningSequence) {
+    public Optional<String> validate(final Integer lengthOfTheWinningSequence) {
         if (lengthOfTheWinningSequence < Const.MINIMAL_LENGTH_OF_WINNING_SEQUENCE) {
             return Optional.of("Length of the winning sequence has to be at least "
                     + Const.MINIMAL_LENGTH_OF_WINNING_SEQUENCE);
